Refactor unit conversion, show individual forecasts as line graphs, remove headwind indicator settings (#137)
* Refactor unit conversion

* ref #136: Fix cloud cover, surface level pressure, sealevel pressure and relative humidity are not included in forecast values

* Unit conversions

* Update pipeline

* Remove apk archival step

* fix #128: Remove absolute displayed wind speed, text on indicator settings, userwindSpeed datatype

* fix #98: Show temperature, precipitation, temperature forecasts as line graphs

* Add WindDirectionAndSpeedDataType

* Disable line graph forecast when using openweathermap

* Fix wind on main menu forecast display

+++ app/src/main/kotlin/de/timklge/karooheadwind/weatherprovider/openweathermap/OpenWeatherMapForecastData.kt
@@ -33,7 +33,7 @@ data class OpenWeatherMapForecastData(
 
         return WeatherData(
             temperature = temp,
-            relativeHumidity = humidity.toDouble(),
+            relativeHumidity = humidity,
             precipitation = rain?.h1 ?: 0.0,
             cloudCover = clouds.toDouble(),
             surfacePressure = pressure.toDouble(),
